Init authorization

+++ routes/campgrounds.js
@@ -1,24 +1,10 @@
 const express = require('express');
 const router = express.Router();
 const catchAsync = require('../utils/catchAsync');
-const { campgroundSchema } = require('../schemas.js');
-const {isLoggedIn} = require('../middleware.js');
-const ExpressError = require('../utils/ExpressError');
+
+const {isLoggedIn, isAuthor, validateCampground} = require('../middleware.js');
+
 const Campground = require('../models/campground');
-const { model } = require('mongoose');
-
-const validateCampground = (req, res, next) => {
-    // validation campground data against schema and grab any errors
-    const {error} = campgroundSchema.validate(req.body);
-
-    // throw errors, otherwise go to next middleware
-    if(error){
-        const msg = error.details.map(el => el.message).join(',')
-        throw new ExpressError(msg, 400);
-    } else {
-        next();
-    }
-}
 
 // route to show index of campgrounds
 router.get('/', catchAsync(async (req, res) =>{
@@ -40,6 +26,9 @@ router.post('/', isLoggedIn, validateCampground, catchAsync(async (req, res, nex
     // grab campground data from request body
     const campground = new Campground(req.body.campground);
 
+    // add user data to campground
+    campground.author = req.user._id;
+
     // save campground data to db and then redirect to page for that campground
     await campground.save();
     req.flash('success', 'Successfully made a new campground.');
@@ -49,8 +38,17 @@ router.post('/', isLoggedIn, validateCampground, catchAsync(async (req, res, nex
 // route for individual campgrounds
 router.get('/:id', catchAsync(async (req, res) => {
     // get campground by database id and populate reviews data
-    const campground = await Campground.findById(req.params.id).populate('reviews');
-
+    const campground = await Campground.findById(req.params.id)
+    .populate({
+        // populate reviews
+        path: 'reviews',
+        populate: {
+            // populate authors for each reviews
+            path:'author'
+        }
+    })
+    // populate author for each campground
+    .populate('author');
     // flash error and redirect to index if campground found
     if(!campground){
         req.flash('error', 'Campground not found');
@@ -61,9 +59,12 @@ router.get('/:id', catchAsync(async (req, res) => {
 }));
 
 // page for editing existing campgrounds
-router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
+    // destructure id from request parameters
+    const {id} = req.params;
+
     // get campground by database id
-    const campground = await Campground.findById(req.params.id);
+    const campground = await Campground.findById(id);
 
     // flash error and redirect to index if campground found
     if(!campground){
@@ -76,7 +77,7 @@ router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
 }));
 
 // route for updating a campground
-router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res) => {
+router.put('/:id', isLoggedIn, isAuthor, validateCampground, catchAsync(async (req, res) => {
     // destructure id from request parameters
     const {id} = req.params;
 
@@ -91,9 +92,16 @@ router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res) =
 }));
 
 // route for deleting campgrounds
-router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
+router.delete('/:id', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
     // destructure id from request parameters
     const {id} = req.params;
+
+    // find campground and check if authro matches logged in user
+    const campground = await Campground.findById(id);
+    if(!campground.author.equals(req.user.id)){
+        req.flash('error', 'You do not have permission to do that');
+        return res.redirect(`/campgrounds/${id}`)
+    }
 
     // find and delete campground from db
     await Campground.findByIdAndDelete(id);
